# Abacws/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.3.0
  title: Abacws Data Visualiser
  description: >
    API for the storage and query of IoT data for the Cardiff University Comptuer Science building (Abacws).
  license:
    name: MIT

tags:
  - name: Advanced Query
    description: >
      Advanced query endpoints for querying the entire dataset

  - name: Devices
    description: >
      Endpoints relating to device info

  - name: Device Data
    description: >
      Endpoints relating to device data entry

  - name: Device History
    description: >
      Endpoints relating to device historical data

servers:
  - url: ./
    description: >
      Relative url
  # - url: https://abacws.ggrainger.uk/api
  #   description: >
  #     Production deployment
  - url: http://localhost:5000/api
    description: >
      Local development server.

paths:
  /devices:
    summary: Device listing endpoint

    get:
      summary: Get a list of all devices
      description: >
        # Device listing endpoint
          Returns a list of all devices.
      tags:
        - Devices
      responses:
        "200":
          description: >
            ##### OK
              List of all devices returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/devices"

        "503":
          $ref: "#/components/responses/databaseFailure"

  /devices/{deviceName}:
    summary: Get information about device
    parameters:
      - $ref: "#/components/parameters/deviceName"

    get:
      summary: Get basic info on a device
      description: >
        # Get information about device
          Returns the basic information about a device.
      tags:
        - Devices
      responses:
        "200":
          description: >
            ##### OK
              Device details returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/device"

  /devices/{deviceName}/data:
    summary: Data interaction endpoint
    parameters:
      - $ref: "#/components/parameters/deviceName"

    get:
      summary: Get current data for a device
      description: >
        # Data interaction endpoint
          Get the last data entry for the given device
      tags:
        - Device Data
      responses:
        "200":
          description: >
            ##### OK
              Device data returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/timestampedData"

        "404":
          $ref: "#/components/responses/deviceNotFound"

        "503":
          $ref: "#/components/responses/databaseFailure"

    put:
      summary: Add data to the given device
      description: >
        # Data interaction endpoint
          Adds data to the given device.
          Data should be provided in JSON format using a PUT request.

          A timestamp is added to the data automatically and is not required.

          Units can optionally be provided by providing an object with a `value` and `units` field instead of just a value:

        ## Without units
            {
              temperature: 20
            }

        ## With units
            {
              temperature: {
                value: 20,
                units: "C"
              }
            }
      tags:
        - Device Data
      security: 
          - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/data"
            example:
              temperature:
                value: 21
                units: °C
      responses:
        "202":
          description: >
            ##### OK
              Data added

        "404":
          $ref: "#/components/responses/deviceNotFound"

        "503":
          $ref: "#/components/responses/databaseFailure"

        "403":
          $ref: "#/components/responses/authFailure"

  /devices/{deviceName}/history:
    summary: Historical data endpoint
    parameters:
      - $ref: "#/components/parameters/deviceName"
      - $ref: "#/components/parameters/fromTime"
      - $ref: "#/components/parameters/toTime"

    get:
      summary: Get historical data
      description: >
        # Historical data endpoint
          Get historical data for the specified device.
          
          Optionally, can provide a "from" or "to" parameter to constrain the timeframe for the request.
      tags:
        - Device History
      responses:
        "200":
          description: >
            ##### OK
              Historical data returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/history"

        "404":
          $ref: "#/components/responses/deviceNotFound"

        "503":
          $ref: "#/components/responses/databaseFailure"

    delete:
      summary: Clear historical data
      description: >
        # Clear historical data
          Deletes all data for the given device
      tags:
        - Device History
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: >
            ##### OK
              Data was clearned

        "404":
          $ref: "#/components/responses/deviceNotFound"

        "503":
          $ref: "#/components/responses/databaseFailure"

        "403":
          $ref: "#/components/responses/authFailure"

  # Advanced query endpoints
  /query:
    summary: Advanced query endpoint
    parameters:
      - $ref: "#/components/parameters/aQueryDeviceNames"
      - $ref: "#/components/parameters/aQueryDeviceTypes"
      - $ref: "#/components/parameters/aQueryDeviceFloors"

    get:
      summary: Advanced query endpoint for device info
      description: >
        # Advanced query endpoint
          Get a filtered list of devices based on provided parameters
      tags:
        - Devices
        - Advanced Query

      responses:
        "200":
          $ref: "#components/responses/aQueryOK"
        "503":
          $ref: "#/components/responses/databaseFailure"

  /query/data:
    summary: Advanced query endpoint
    parameters:
      - $ref: "#/components/parameters/aQueryDeviceNames"
      - $ref: "#/components/parameters/aQueryDeviceTypes"
      - $ref: "#/components/parameters/aQueryDeviceFloors"
      - $ref: "#/components/parameters/aQueryDeviceHasFields"

    get:
      summary: Advanced query endpoint for device data
      description: >
        # Advanced query endpoint
          Get a filtered list of devices and their last data entry based on provided parameters.

          NOTE: Devices with no data matching the query will not be returned.
      tags:
        - Device Data
        - Advanced Query

      responses:
        "200":
          $ref: "#components/responses/aQueryOK"
        "503":
          $ref: "#/components/responses/databaseFailure"

  /query/history:
    summary: Advanced query endpoint
    parameters:
      - $ref: "#/components/parameters/aQueryDeviceNames"
      - $ref: "#/components/parameters/aQueryDeviceTypes"
      - $ref: "#/components/parameters/aQueryDeviceFloors"
      - $ref: "#/components/parameters/aQueryDeviceHasFields"
      - $ref: "#/components/parameters/fromTime"
      - $ref: "#/components/parameters/toTime"

    get:
      summary: Advanced query endpoint for device history
      description: >
        # Advanced query endpoint
          Get a filtered list of devices and associated historical data based on the provided parameters

          NOTE: Devices with no data matching the query will not be returned.
      tags:
        - Device History
        - Advanced Query

      responses:
        "200":
          $ref: "#components/responses/aQueryOK"
        "503":
          $ref: "#/components/responses/databaseFailure"


components:
  responses:
    deviceNotFound:
      description: >
        ##### Device not found
          The requested device does not exist in the database.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
          example:
            error: Device does not exist

    databaseFailure:
      description: >
        ##### Unable to access database
          The API was unable to read/write from the database.
          This may mean that the database is unreachable or that the api lacks the permissions to read/write from/to it.
          Check your logs for more information.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/error"
          example:
            error: Unable to access database

    authFailure:
      description: >
        ##### Auth failure
          The authorisation provided was insufficient for this endpoint.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/error"
          example:
            error: You do not have permission to access this resource

    # Advaned query responses
    aQueryOK:
      description: >
        ##### OK
          A list of the devices matching the provided parameters was returned.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/devices"


  parameters:
    deviceName:
      in: path
      name: deviceName
      required: true
      description: >
        Name of the the device you wish to query
      schema:
        type: string
        example: Room-0.01

    fromTime:
      in: query
      name: from
      required: false
      description: >
        Defines the start of a date range for history requests.
          - If specified will only allow data points with a time larger than the provided value.
          - Should be formatted in UNIX time.
      schema:
        type: number

    toTime:
      in: query
      name: to
      required: false
      description: >
        Defines the end of a date range for history requests.
          - If specified will only allow data points with a time smaller than the provided value.
          - Should be formatted in UNIX time.
      schema:
        type: number

    # Advanced query parameters
    aQueryDeviceNames:
      name: name
      in: query
      required: false
      description: >
        The names of the devices we wish to query.
          - Can be a single value, or a list of comma-delimited values
      schema:
        type: string
        examples:
          - room1
          - room1,room2,room3

    aQueryDeviceTypes:
      name: type
      in: query
      required: false
      description: >
        Types of devices we wish to query.
          - Can be a single value, or a list of comma-delimited values
      schema:
        type: string
        examples:
          - lecture
          - lecture,office,meeting

    aQueryDeviceFloors:
      name: floor
      in: query
      required: false
      description: >
        Floors on which the devices we wish to query are
          - Can be a single value, or a list of comma-delimited values
      schema:
        type: string
        examples:
          - 1
          - 1,2,3

    aQueryDeviceHasFields:
      name: has
      in: query
      required: false
      description: >
        The fields we want the returned data to contain.
          - Devices and data not containing these fields will not be returned.
          - Can be a single value, or a list of comma-delimited values
      schema:
        type: string
        examples:
          - temperature
          - temperature,humidity,light level


  schemas:
    device:
      type: object
      description: >
        A device represents a real world IoT device with:
            A known location
            A unique name which can be used to identify it.
      required:
        - name
        - type
        - position
      properties:
        name:
          type: string
          description: >
            Unique name of the device
          examples:
            - Room-0.01
        type:
          type: string
          description: >
            The type of device
          examples:
            - lecture
            - office
        position:
          $ref: "#/components/schemas/position"
        data:
          $ref: "#/components/schemas/timestampedData"
        history:
          $ref: "#/components/schemas/history"

    devices:
      description: >
        A list of devices
      type: array
      items:
        $ref: "#/components/schemas/device"

    position:
      type: object
      description: >
        The absolute position of a device in the world
      properties:
        x:
          type: number
        y:
          type: number
        z:
          type: number

    timestamp:
      type: integer
      description: UNIX timestamp
      examples:
        - 1647527656

    data:
      type: object
      description: >
        Datapoint provided by a device, can have any number of fields.
        For example, a device could provide a temperature value.
      examples:
        - temperature:
            value: 20
            units: °C
          boolean value:
            value: true

    timestampedData:
      type: object
      description: >
        Datapoint provided by a device, can have any number of fields.
        Will always include a timestamp field indiciating when this data was provided in UNIX time.
      allOf:
        - $ref: "#/components/schemas/data"
        - type: object
          properties:
            timestamp:
              $ref: "#/components/schemas/timestamp"

          examples:
            - timestamp: 1647527656
              temperature:
                value: 20
                units: °C
              boolean value:
                value: true

    history:
      type: array
      minItems: 0
      description: >
        Array of historical, timestamped data.
      items:
        $ref: "#/components/schemas/timestampedData"

    error:
      type: object
      description: >
        An error of some description was encountered by the API.
        A summary of this error will be provided as part of the error field.
      properties:
        error:
          type: string
          examples:
            - Not Found
            - Could not read/write from database


  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key